Single TPR parse, directory-based trajectory CLI, per-calculator timing, float32 aim
FullSystemReader: ReadTopology parses TPR once — atom ranges, bonded
parameters, and protein construction from one read_tpx_state. Previously
three independent parses. BuildProtein() method replaces standalone
BuildProteinFromTpr for trajectory path. detail::TprData holds the
gmx_mtop_t without GROMACS headers in the .h.

GromacsRunContext::Build takes (reader, edr_path), no tpr_path.
Bonded params from reader's stored parse.

--trajectory DIR: single directory argument, derives md.tpr/md.xtc/md.edr.
All three required. No --tpr/--xtc/--edr flags. No file discovery.

BuildFromTrajectory(dir_path): also fixes pre-existing bug where
charges from BuildResult were dropped (moved protein but not charges).

OperationRunner: TimedAttach wraps every calculator Compute in a
Scope — per-calculator [BEGIN]/[END] with elapsed ms over UDP.

AIMNet2 aim embedding: float32 through the full path (ConformationAtom,
NPY via WriteFloat32, H5 via WriteRawFloat). Was float64 upshift from
native torch — wasted 0.3 GB per protein.

Old fleet PDB test removed. Tests point at walker_0/ which has the
production directory layout (md.tpr, md.xtc, md.edr).

Validated: 1CBH_192 analysis H5 adversarial check passed. NPY-vs-H5
bit-identical. No NaN/Inf. Physics signs correct. MolProbity 0 twisted.
Fleet extraction running on 10 calibration proteins.

+++ src/AIMNet2Result.cpp
@@ -299,8 +299,8 @@ std::unique_ptr<AIMNet2Result> AIMNet2Result::Compute(
 
     {
         auto aim_gpu = output_dict.at("aim").toTensor();
-        auto aim_cpu = aim_gpu.to(torch::kCPU, torch::kFloat64);
-        auto aim_acc = aim_cpu.accessor<double, 2>();
+        auto aim_cpu = aim_gpu.to(torch::kCPU, torch::kFloat32);
+        auto aim_acc = aim_cpu.accessor<float, 2>();
 
         int64_t model_dims = aim_cpu.size(1);
         if (model_dims != static_cast<int64_t>(AIMNET2_AIM_DIMS)) {
@@ -485,13 +485,13 @@ int AIMNet2Result::WriteFeatures(
         files_written++;
     }
 
-    // aimnet2_aim.npy — (N, AIMNET2_AIM_DIMS) float64
+    // aimnet2_aim.npy — (N, AIMNET2_AIM_DIMS) float32 (native torch precision)
     {
-        std::vector<double> data(N * AIMNET2_AIM_DIMS);
+        std::vector<float> data(N * AIMNET2_AIM_DIMS);
         for (size_t i = 0; i < N; ++i)
             for (size_t d = 0; d < AIMNET2_AIM_DIMS; ++d)
                 data[i * AIMNET2_AIM_DIMS + d] = conf.AtomAt(i).aimnet2_aim[d];
-        NpyWriter::WriteFloat64(output_dir + "/aimnet2_aim.npy",
+        NpyWriter::WriteFloat32(output_dir + "/aimnet2_aim.npy",
                                 data.data(), N, AIMNET2_AIM_DIMS);
         files_written++;
     }
